Version 1.3
- Change version to 1.3.
- Updated for Xcode 11.1 and macOS 10.15 Catalina.
- Minimum macOS version supported is 10.6.
- 32-bit processors are no longer supported.
- Growl is no longer supported.

+++ StartTimerDialogController.m
@@ -3,7 +3,7 @@
 //  MenuTimer
 //
 //  Created by Kristopher Johnson on 3/29/09.
-//  Copyright 2009 Capable Hands Technologies, Inc.. All rights reserved.
+//  Copyright © 2009,2015,2019 Kristopher Johnson
 //
 //  This file is part of Menubar Countdown.
 //
@@ -47,9 +47,9 @@
 
 - (NSTimeInterval)timerInterval {
     NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
-    int timerHours = [defaults integerForKey:UserDefaultsTimerHoursKey];
-    int timerMinutes = [defaults integerForKey:UserDefaultsTimerMinutesKey];
-    int timerSeconds = [defaults integerForKey:UserDefaultsTimerSecondsKey];
+    NSInteger timerHours = [defaults integerForKey:UserDefaultsTimerHoursKey];
+    NSInteger timerMinutes = [defaults integerForKey:UserDefaultsTimerMinutesKey];
+    NSInteger timerSeconds = [defaults integerForKey:UserDefaultsTimerSecondsKey];
     return (timerHours * 3600) + (timerMinutes * 60) + timerSeconds;
 }
